커뮤니티 게시글, 댓글, 추천/비추천, 스크랩 리팩토링 및 해시태그 기능추가

+++ api-module/src/main/java/org/devridge/api/domain/community/service/CommunityScrapService.java
@@ -5,13 +5,15 @@ import lombok.RequiredArgsConstructor;
 import org.devridge.api.domain.community.entity.Community;
 import org.devridge.api.domain.community.entity.CommunityScrap;
 import org.devridge.api.domain.community.entity.id.CommunityScrapId;
+import org.devridge.api.domain.community.mapper.CommunityScrapMapper;
 import org.devridge.api.domain.community.repository.CommunityRepository;
 import org.devridge.api.domain.community.repository.CommunityScrapRepository;
 import org.devridge.api.domain.member.entity.Member;
 import org.devridge.api.domain.member.repository.MemberRepository;
 import org.devridge.api.util.SecurityContextHolderUtil;
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
@@ -20,49 +22,34 @@ public class CommunityScrapService {
     private final CommunityScrapRepository communityScrapRepository;
     private final MemberRepository memberRepository;
     private final CommunityRepository communityRepository;
+    private final CommunityScrapMapper communityScrapMapper;
 
+    @Transactional
     public void createScrap(Long communityId) {
-        Long writeMemberId = SecurityContextHolderUtil.getMemberId();
-        Member member = getMemberById(writeMemberId);
+        Long accessMemberId = SecurityContextHolderUtil.getMemberId();
+        Member member = getMemberById(accessMemberId);
         Community community = getCommunityById(communityId);
-        CommunityScrapId communityScrapId = new CommunityScrapId(writeMemberId, communityId);
+        CommunityScrapId communityScrapId = new CommunityScrapId(accessMemberId, communityId);
 
-        communityScrapRepository.findById(communityScrapId)
-            .ifPresent(cs -> {
-                throw new DataIntegrityViolationException("이미 존재");
-            });
+        if (!accessMemberId.equals(community.getMember().getId())) {
+            throw new AccessDeniedException("거부된 접근입니다.");
+        }
 
-        if (communityScrapRepository.findById(communityScrapId).isEmpty()) { // 없나확인
-            Long countResult = communityScrapRepository.checkSoftDelete(communityId, writeMemberId); // 0이외의 값은 가짜없음
-            boolean isSoftDeleted = countResult != 0L;
+        communityScrapRepository.findById(communityScrapId).ifPresentOrElse(
+            result -> {
+                if (result.getIsDeleted()) {
+                    communityScrapRepository.restoreById(communityScrapId);
+                }
 
-            //가짜없음
-            //reCreateScrap쿼리하기
-            if (isSoftDeleted) {  // 참이면 가짜없음임 즉 실행되면
-                communityScrapRepository.reCreateScrap(communityId, writeMemberId);
-            }
-
-            //  진짜없음
-            //  새로만들기
-            if (!isSoftDeleted) {
-                CommunityScrap communityScrap = CommunityScrap.builder()
-                    .community(community)
-                    .member(member)
-                    .build();
+                if (!result.getIsDeleted()) {
+                    communityScrapRepository.deleteById(communityScrapId);
+                }
+            },
+            () -> {
+                CommunityScrap communityScrap = communityScrapMapper.toCommunityScrap(community, member);
                 communityScrapRepository.save(communityScrap);
             }
-        }
-    }
-
-    public void deleteScrap(Long communityId) {
-        Long writeMemberId = SecurityContextHolderUtil.getMemberId();
-        Member member = getMemberById(writeMemberId);
-        getCommunityById(communityId);
-
-        CommunityScrapId communityScrapId = new CommunityScrapId(writeMemberId, communityId);
-        communityScrapRepository.findById(communityScrapId)
-            .orElseThrow(() -> new EntityNotFoundException("스크랩이 존재하지 않습니다."));
-        communityScrapRepository.deleteById(communityScrapId);
+        );
     }
 
     private Member getMemberById(Long memberId) {
